Avoid segfaults in tests with MinGW-w64

+++ testsuite/adapters/schwartz_adapter_every_sorter.cpp
@@ -35,7 +35,8 @@
 
 namespace
 {
-    template<typename T=double>
+    // double segfaults with MinGW-w64 64 bits in Release mode
+    template<typename T=float>
     struct wrapper
     {
         T value;
